ion: Fix possible errors in MSM ION driver
PVS-Studio:
V547 Expression 'val < 0' is always false. Unsigned type value is never < 0

+++ drivers/gpu/ion/msm/msm_ion.c
@@ -424,7 +424,7 @@ static int msm_ion_get_heap_size(struct device_node *node,
 
 		if (!ret && memory_name_prop) {
 			val = msm_get_memory_type_from_name(memory_name_prop);
-			if (val < 0) {
+			if ((int)val < 0) {
 				ret = -EINVAL;
 				goto out;
 			}
